requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
seaborn>=0.13.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
fpdf2>=2.7.0
kaleido>=0.2.1
